python/ch07/llamaindex/rag_agent_with_context_retrieval/requirements.txt
llama-index
llama-index-core
llama-index-agent-openai
llama-index-llms-openai
llama-index-embeddings-openai
openai
python-dotenv